+++ immutable-arrays/transformations-to-standard-collections/src/main/kotlin/com/danrusu/pods4k/immutableArrays/TransformationsToMap.kt
@@ -1,0 +1,10 @@
+// Auto-generated file. DO NOT EDIT!
+package com.danrusu.pods4k.immutableArrays
+
+import kotlin.Pair
+import kotlin.collections.Map
+
+/**
+ * See [Array.toMap]
+ */
+public fun <K, V> ImmutableArray<Pair<K, V>>.toMap(): Map<K, V> = asList().toMap()